backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai>=0.3.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==0.8.11
python-pptx==0.6.21
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.13
aiofiles==23.2.1
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
